[TSK-001] Implementing gemini model execution via aistudio client and pdf beautified generation by using libraries merging html and css definition for styling the content

+++ src/main/java/es/NTTEnterprise/RIntellix/ms_reporting/utils/ReportConstants.java
@@ -2,7 +2,8 @@ package es.NTTEnterprise.RIntellix.ms_reporting.utils;
 
 /**
  * Centralized constants for the ms-reporting microservice.
- * This class contains reusable and consistent constants to avoid hardcoding values.
+ * This class contains reusable and consistent constants to avoid hardcoding
+ * values.
  * 
  * @author Lucía Fernández Mancebo
  * @Date 29-06-2026
@@ -29,7 +30,7 @@ public final class ReportConstants {
     // Gemini API Configurations
     public static final String GEMINI_GENERATE_CONTENT_PATH = "/v1beta/models/{model}:generateContent";
     public static final String RESPONSE_MIME_JSON = "application/json";
-    public static final double GEMINI_TEMPERATURE = 0.2;
+    public static final Float GEMINI_TEMPERATURE = 0.2F;
     public static final String GEMINI_USER_ROLE = "user";
     public static final String GEMINI_API_KEY_QUERY_PARAM = "key";
 
@@ -39,15 +40,15 @@ public final class ReportConstants {
     public static final String PDF_SECTION_ANALYSIS = "Análisis de riesgo";
     public static final String PDF_SECTION_FACTORS = "Factores de riesgo";
     public static final String PDF_SECTION_RECOMMENDATIONS = "Recomendaciones";
-    
+
     public static final String PDF_TABLE_HEADER_FACTOR = "Factor";
     public static final String PDF_TABLE_HEADER_SEVERITY = "Severidad";
     public static final String PDF_TABLE_HEADER_DESCRIPTION = "Descripción";
-    
+
     public static final String PDF_HYPHEN = "-";
     public static final String PDF_BULLET_PREFIX = "• ";
     public static final String PDF_FILE_NAME_PATTERN = "report_%s_%s.pdf";
-    
+
     public static final String PDF_META_REQUEST_PREFIX = "Request: ";
     public static final String PDF_META_SCORING_SEPARATOR = "  |  Scoring: ";
     public static final String PDF_META_MODEL_SEPARATOR = "  |  Modelo IA: ";
